*** empty log message ***
svn path=/trunk/boinc/; revision=3863

+++ sched/validate_util.h
@@ -35,7 +35,7 @@ int generic_check_set(
     init_result_f init_result_f,
     check_pair_with_data_f check_pair_with_data_f,
     cleanup_result_f cleanup_result_f,
-    int min_valid);
+    size_t min_valid);
 
 int generic_check_set_majority(
     std::vector<RESULT>& results, int& canonicalid, double& credit,
